+++ R/optim_switch.R
@@ -32,24 +32,32 @@ optim_switch <- function(...) {
 
         # Save results even in case parameter estimation crash
         res$obs_var_list <- .croptEnv$obs_var_list
-        rm("obs_var_list", envir = .croptEnv)
+        if (exists(".croptEnv$obs_var_list")) {
+          rm("obs_var_list", envir = .croptEnv)
+        }
 
         if (arguments$crit_options$info_level >= 1) {
           res$params_and_crit <- dplyr::bind_rows(.croptEnv$params_and_crit)
-          rm("params_and_crit", envir = .croptEnv)
+          if (exists(".croptEnv$params_and_crit")) {
+            rm("params_and_crit", envir = .croptEnv)
+          }
         }
         if (arguments$crit_options$info_level >= 2) {
           res$sim_intersect <- .croptEnv$sim_intersect
-          rm("sim_intersect", envir = .croptEnv)
+          if (exists(".croptEnv$sim_intersect")) {
+            rm("sim_intersect", envir = .croptEnv)
+          }
         }
         if (arguments$crit_options$info_level >= 3) {
           res$obs_intersect <- .croptEnv$obs_intersect
-          rm("obs_intersect", envir = .croptEnv)
+          if (exists(".croptEnv$obs_intersect")) {
+            rm("obs_intersect", envir = .croptEnv)
+          }
         }
         if (arguments$crit_options$info_level >= 4) {
           res$sim <- .croptEnv$sim
           res$sim_transformed <- .croptEnv$sim_transformed
-          if (!is.null(res$sim_transformed)) {
+          if (exists(".croptEnv$sim_transformed")) {
             rm("sim_transformed", envir = .croptEnv)
           }
         }